Demo doc -> ord mapping using aggregation
We add country documents, corresponding to refinements and rank them
using aggregations over the point of itnerest and data from the
country docs.

This gives us a good opportunity to showcase how we can use faceting
to map a values source over docs (the conutry docs) to one over
ordinals (the ordinals corresponding to the countries).